% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adipose_Subcutaneous.R
\docType{data}
\name{Adipose_Subcutaneous}
\alias{Adipose_Subcutaneous}
\title{Matrix of expression data of Adipose...Subcutaneous from gtex}
\format{
A [[matrix]] object
}
\source{
[Adipose_Subcutaneous](https://gtexportal.org/home/datasets)
}
\usage{
Adipose_Subcutaneous
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
