% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adipose_Visceral_Omentum.R
\docType{data}
\name{Adipose_Visceral_Omentum}
\alias{Adipose_Visceral_Omentum}
\title{Matrix of expression data of Adipose...Visceral..Omentum. from gtex}
\format{
A [[matrix]] object
}
\source{
[Adipose_Visceral_Omentum](https://gtexportal.org/home/datasets)
}
\usage{
Adipose_Visceral_Omentum
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
