% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrenal_Gland.R
\docType{data}
\name{Adrenal_Gland}
\alias{Adrenal_Gland}
\title{Matrix of expression data of Adrenal.Gland from gtex}
\format{
A [[matrix]] object
}
\source{
[Adrenal_Gland](https://gtexportal.org/home/datasets)
}
\usage{
Adrenal_Gland
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
