% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artery_Aorta.R
\docType{data}
\name{Artery_Aorta}
\alias{Artery_Aorta}
\title{Matrix of expression data of Artery...Aorta from gtex}
\format{
A [[matrix]] object
}
\source{
[Artery_Aorta](https://gtexportal.org/home/datasets)
}
\usage{
Artery_Aorta
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
