% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artery_Coronary.R
\docType{data}
\name{Artery_Coronary}
\alias{Artery_Coronary}
\title{Matrix of expression data of Artery...Coronary from gtex}
\format{
A [[matrix]] object
}
\source{
[Artery_Coronary](https://gtexportal.org/home/datasets)
}
\usage{
Artery_Coronary
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
