% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artery_Tibial.R
\docType{data}
\name{Artery_Tibial}
\alias{Artery_Tibial}
\title{Matrix of expression data of Artery...Tibial from gtex}
\format{
A [[matrix]] object
}
\source{
[Artery_Tibial](https://gtexportal.org/home/datasets)
}
\usage{
Artery_Tibial
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
