% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bladder.R
\docType{data}
\name{Bladder}
\alias{Bladder}
\title{Matrix of expression data of Bladder from gtex}
\format{
A [[matrix]] object
}
\source{
[Bladder](https://gtexportal.org/home/datasets)
}
\usage{
Bladder
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
