% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Amygdala.R
\docType{data}
\name{Brain_Amygdala}
\alias{Brain_Amygdala}
\title{Matrix of expression data of Brain...Amygdala from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Amygdala](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Amygdala
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
