% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Anterior_cingulate_cortex_BA24.R
\docType{data}
\name{Brain_Anterior_cingulate_cortex_BA24}
\alias{Brain_Anterior_cingulate_cortex_BA24}
\title{Matrix of expression data of Brain_Anterior.cingulate.cortex..BA24. from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Anterior_cingulate_cortex_BA24](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Anterior_cingulate_cortex_BA24
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
