% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Cerebellar_Hemisphere.R
\docType{data}
\name{Brain_Cerebellar_Hemisphere}
\alias{Brain_Cerebellar_Hemisphere}
\title{Matrix of expression data of Brain...Cerebellar.Hemisphere from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Cerebellar_Hemisphere](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Cerebellar_Hemisphere
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
