% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Cerebellum.R
\docType{data}
\name{Brain_Cerebellum}
\alias{Brain_Cerebellum}
\title{Matrix of expression data of Brain...Cerebellum from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Cerebellum](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Cerebellum
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
