% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Cortex.R
\docType{data}
\name{Brain_Cortex}
\alias{Brain_Cortex}
\title{Matrix of expression data of Brain...Cortex from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Cortex](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Cortex
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
