% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Frontal_Cortex_BA9.R
\docType{data}
\name{Brain_Frontal_Cortex_BA9}
\alias{Brain_Frontal_Cortex_BA9}
\title{Matrix of expression data of Brain...Frontal.Cortex..BA9. from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Frontal_Cortex_BA9](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Frontal_Cortex_BA9
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
