% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Hippocampus.R
\docType{data}
\name{Brain_Hippocampus}
\alias{Brain_Hippocampus}
\title{Matrix of expression data of Brain...Hippocampus from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Hippocampus](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Hippocampus
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
