% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Hypothalamus.R
\docType{data}
\name{Brain_Hypothalamus}
\alias{Brain_Hypothalamus}
\title{Matrix of expression data of Brain...Hypothalamus from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Hypothalamus](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Hypothalamus
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
