% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Nucleus_accumbens_basal_ganglia.R
\docType{data}
\name{Brain_Nucleus_accumbens_basal_ganglia}
\alias{Brain_Nucleus_accumbens_basal_ganglia}
\title{Matrix of expression data of Brain...Nucleus.accumbens..basal.ganglia. from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Nucleus_accumbens_basal_ganglia](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Nucleus_accumbens_basal_ganglia
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
