% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Spinal_cord_cervical_c_1.R
\docType{data}
\name{Brain_Spinal_cord_cervical_c_1}
\alias{Brain_Spinal_cord_cervical_c_1}
\title{Matrix of expression data of Brain...Spinal.cord..cervical.c.1. from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Spinal_cord_cervical_c_1](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Spinal_cord_cervical_c_1
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
