% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brain_Substantia_nigra.R
\docType{data}
\name{Brain_Substantia_nigra}
\alias{Brain_Substantia_nigra}
\title{Matrix of expression data of Brain_Substantia.nigra from gtex}
\format{
A [[matrix]] object
}
\source{
[Brain_Substantia_nigra](https://gtexportal.org/home/datasets)
}
\usage{
Brain_Substantia_nigra
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
