% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breast_Mammary_Tissue.R
\docType{data}
\name{Breast_Mammary_Tissue}
\alias{Breast_Mammary_Tissue}
\title{Matrix of expression data of Breast_Mammary.Tissue from gtex}
\format{
A [[matrix]] object
}
\source{
[Breast_Mammary_Tissue](https://gtexportal.org/home/datasets)
}
\usage{
Breast_Mammary_Tissue
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
