% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells_Cultured_fibroblasts.R
\docType{data}
\name{Cells_Cultured_fibroblasts}
\alias{Cells_Cultured_fibroblasts}
\title{Matrix of expression data of Cells...Cultured.fibroblasts from gtex}
\format{
A [[matrix]] object
}
\source{
[Cells_Cultured_fibroblasts](https://gtexportal.org/home/datasets)
}
\usage{
Cells_Cultured_fibroblasts
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
