% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells_EBV_transformed_lymphocytes.R
\docType{data}
\name{Cells_EBV_transformed_lymphocytes}
\alias{Cells_EBV_transformed_lymphocytes}
\title{Matrix of expression data of Cells...EBV.transformed.lymphocytes from gtex}
\format{
A [[matrix]] object
}
\source{
[Cells_EBV_transformed_lymphocytes](https://gtexportal.org/home/datasets)
}
\usage{
Cells_EBV_transformed_lymphocytes
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
