% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervix_Ectocervix.R
\docType{data}
\name{Cervix_Ectocervix}
\alias{Cervix_Ectocervix}
\title{Matrix of expression data of Cervix...Ectocervix from gtex}
\format{
A [[matrix]] object
}
\source{
[Cervix_Ectocervix](https://gtexportal.org/home/datasets)
}
\usage{
Cervix_Ectocervix
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
