% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervix_Endocervix.R
\docType{data}
\name{Cervix_Endocervix}
\alias{Cervix_Endocervix}
\title{Matrix of expression data of Cervix...Endocervix from gtex}
\format{
A [[matrix]] object
}
\source{
[Cervix_Endocervix](https://gtexportal.org/home/datasets)
}
\usage{
Cervix_Endocervix
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
