% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatin1.R
\docType{data}
\name{Chromatin1}
\alias{Chromatin1}
\title{GRanges object of chromatin states of whole genome}
\format{
A [[Granges]] object
}
\source{
[Chromatin1](http://www.roadmapepigenomics.org/)
}
\usage{
Chromatin1
}
\description{
Roadmap epigenomics project
}
\keyword{datasets}
