% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatin2.R
\docType{data}
\name{Chromatin2}
\alias{Chromatin2}
\title{GRanges object of chromatin states of whole genome}
\format{
A [[Granges]] object
}
\source{
[Chromatin2](http://www.roadmapepigenomics.org/)
}
\usage{
Chromatin2
}
\description{
Roadmap epigenomics project
}
\keyword{datasets}
