% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatin3.R
\docType{data}
\name{Chromatin3}
\alias{Chromatin3}
\title{GRanges object of chromatin states of whole genome}
\format{
A [[Granges]] object
}
\source{
[Chromatin3](http://www.roadmapepigenomics.org/)
}
\usage{
Chromatin3
}
\description{
Roadmap epigenomics project
}
\keyword{datasets}
