% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromatinGroups.R
\docType{data}
\name{ChromatinGroups}
\alias{ChromatinGroups}
\title{List of grouped tissues}
\format{
A [[List]] object
}
\source{
[ChromatinGroups]
}
\usage{
ChromatinGroups
}
\description{
Chromatin tissues grouped
}
\keyword{datasets}
