% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromatinInteractions.R
\docType{data}
\name{ChromatinInteractions}
\alias{ChromatinInteractions}
\title{GRanges object of chromatin interactions of whole genome}
\format{
A [[Granges]] object
}
\source{
[ChromatinInteractions](https://4dgenome.research.chop.edu/)
}
\usage{
ChromatinInteractions
}
\description{
4D Genome
}
\keyword{datasets}
