% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colon_Sigmoid.R
\docType{data}
\name{Colon_Sigmoid}
\alias{Colon_Sigmoid}
\title{Matrix of expression data of Colon...Sigmoid from gtex}
\format{
A [[matrix]] object
}
\source{
[Colon_Sigmoid](https://gtexportal.org/home/datasets)
}
\usage{
Colon_Sigmoid
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
