% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colon_Transverse.R
\docType{data}
\name{Colon_Transverse}
\alias{Colon_Transverse}
\title{Matrix of expression data of Colon...Transverse from gtex}
\format{
A [[matrix]] object
}
\source{
[Colon_Transverse](https://gtexportal.org/home/datasets)
}
\usage{
Colon_Transverse
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
