% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Esophagus_Gastroesophageal_Junction.R
\docType{data}
\name{Esophagus_Gastroesophageal_Junction}
\alias{Esophagus_Gastroesophageal_Junction}
\title{Matrix of expression data of Esophagus...Gastroesophageal.Junction from gtex}
\format{
A [[matrix]] object
}
\source{
[Esophagus_Gastroesophageal_Junction](https://gtexportal.org/home/datasets)
}
\usage{
Esophagus_Gastroesophageal_Junction
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
