% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Esophagus_Mucosa.R
\docType{data}
\name{Esophagus_Mucosa}
\alias{Esophagus_Mucosa}
\title{Matrix of expression data of Esophagus...Mucosa from gtex}
\format{
A [[matrix]] object
}
\source{
[Esophagus_Mucosa](https://gtexportal.org/home/datasets)
}
\usage{
Esophagus_Mucosa
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
