% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Esophagus_Muscularis.R
\docType{data}
\name{Esophagus_Muscularis}
\alias{Esophagus_Muscularis}
\title{Matrix of expression data of Esophagus...Muscularis from gtex}
\format{
A [[matrix]] object
}
\source{
[Esophagus_Muscularis](https://gtexportal.org/home/datasets)
}
\usage{
Esophagus_Muscularis
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
