% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fallopian_Tube.R
\docType{data}
\name{Fallopian_Tube}
\alias{Fallopian_Tube}
\title{Matrix of expression data of Fallopian.Tube from gtex}
\format{
A [[matrix]] object
}
\source{
[Fallopian_Tube](https://gtexportal.org/home/datasets)
}
\usage{
Fallopian_Tube
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
