% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtexTissuesV8.R
\docType{data}
\name{gtexTissuesV8}
\alias{gtexTissuesV8}
\title{Vector of gtex tissues V8}
\format{
A [[vector]] object
}
\source{
[Bladder](https://gtexportal.org/home/datasets)
}
\usage{
gtexTissuesV8
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
