% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heart_Atrial_Appendage.R
\docType{data}
\name{Heart_Atrial_Appendage}
\alias{Heart_Atrial_Appendage}
\title{Matrix of expression data of Heart...Atrial.Appendage from gtex}
\format{
A [[matrix]] object
}
\source{
[Heart_Atrial_Appendage](https://gtexportal.org/home/datasets)
}
\usage{
Heart_Atrial_Appendage
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
