% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heart_Left_Ventricle.R
\docType{data}
\name{Heart_Left_Ventricle}
\alias{Heart_Left_Ventricle}
\title{Matrix of expression data of Heart...Left.Ventricle from gtex}
\format{
A [[matrix]] object
}
\source{
[Heart_Left_Ventricle](https://gtexportal.org/home/datasets)
}
\usage{
Heart_Left_Ventricle
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
