% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kidney_Cortex.R
\docType{data}
\name{Kidney_Cortex}
\alias{Kidney_Cortex}
\title{Matrix of expression data of Kidney...Cortex from gtex}
\format{
A [[matrix]] object
}
\source{
[Kidney_Cortex](https://gtexportal.org/home/datasets)
}
\usage{
Kidney_Cortex
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
