% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kidney_Medulla.R
\docType{data}
\name{Kidney_Medulla}
\alias{Kidney_Medulla}
\title{Matrix of expression data of Kidney...Medulla from gtex}
\format{
A [[matrix]] object
}
\source{
[Kidney_Medulla](https://gtexportal.org/home/datasets)
}
\usage{
Kidney_Medulla
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
