% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liver.R
\docType{data}
\name{Liver}
\alias{Liver}
\title{Matrix of expression data of Liver from gtex}
\format{
A [[matrix]] object
}
\source{
[Liver](https://gtexportal.org/home/datasets)
}
\usage{
Liver
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
