% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lQTLs.R
\docType{data}
\name{lqtls}
\alias{lqtls}
\title{Data frame of lQTLs}
\format{
A [[data.frame]] object
}
\source{
[lQTLs](see description)
}
\usage{
lqtls
}
\description{
Author: R. Tabassum
Title: Genetic architecture of human plasma lipidome and its link to cardiovascular disease
Journal: NATURE COMMUNICATIONS | https://doi.org/10.1038/s41467-019-11954-8
}
\keyword{datasets}
