% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lung.R
\docType{data}
\name{Lung}
\alias{Lung}
\title{Matrix of expression data of Lung from gtex}
\format{
A [[matrix]] object
}
\source{
[Lung](https://gtexportal.org/home/datasets)
}
\usage{
Lung
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
