% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meQTLs.R
\docType{data}
\name{meQTLs}
\alias{meQTLs}
\title{Data frame of meQTLs}
\format{
A [[data.frame]] object
}
\source{
[meQTLs](https://genenetwork.nl/biosqtlbrowser/)
}
\usage{
meQTLs
}
\description{
Biosqtlbrowser
}
\keyword{datasets}
