% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minor_Salivary_Gland.R
\docType{data}
\name{Minor_Salivary_Gland}
\alias{Minor_Salivary_Gland}
\title{Matrix of expression data of Minor.Salivary.Gland from gtex}
\format{
A [[matrix]] object
}
\source{
[Minor_Salivary_Gland](https://gtexportal.org/home/datasets)
}
\usage{
Minor_Salivary_Gland
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
