% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miQTLexperiment.R
\docType{data}
\name{miQTLexperiment}
\alias{miQTLexperiment}
\title{Data table of experimental determined miQTLs}
\format{
A [[data.table]] object
}
\source{
[miQTLexperiment](see description)
}
\usage{
miQTLexperiment
}
\description{
Huan, T. et al. Genome-wide identification of microRNA expression quantitative trait loci. Nat. Commun. 6:6601 doi: 10.1038/ncomms7601 (2015).
}
\details{
Borel, C., et al., Identification of cis-and trans-regulatory variation modulating microRNA expression levels in human fibroblasts. Genome research, 2011. 21(1): p. 68-73.

Gamazon, E.R., et al., Genetic architecture of microRNA expression: implications for the transcriptome and complex traits. The American Journal of Human Genetics, 2012. 90(6): p. 1046-1063
}
\keyword{datasets}
