% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miQTLpredict.R
\docType{data}
\name{miQTLpredict}
\alias{miQTLpredict}
\title{Data table of predicted miQTLs}
\format{
A [[data.table]] object
}
\source{
[miQTLpredict](http://bioinfo.bjmu.edu.cn/mirsnp/download/)
}
\usage{
miQTLpredict
}
\description{
Liu, C., et al., MirSNP, a database of polymorphisms altering miRNA target sites, identifies miRNA-related SNPs in GWAS SNPs and eQTLs. BMC genomics, 2012. 13(1): p. 661
}
\keyword{datasets}
