% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqtls_LC.R
\docType{data}
\name{mqtls_LC}
\alias{mqtls_LC}
\title{Data frame of lQTLs}
\format{
A [[data.frame]] object
}
\source{
[mQTLs](see description)
}
\usage{
mqtls_LC
}
\description{
Author: E. Rhee
Title: A Genome-wide Association Study of the Human Metabolome in a Community-Based Cohort
Journal: Cell Metab DOI:https://doi.org/10.1016/j.cmet.2013.06.013
GWAS Data for All 217 Metabolites Measured by Our Platform, Including All Loci with P < 1.0 × 10−3
}
\keyword{datasets}
