% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mQTLsNG.R
\docType{data}
\name{mqtls_NG}
\alias{mqtls_NG}
\title{Data frame of lQTLs}
\format{
A [[data.frame]] object
}
\source{
[mQTLs](see description)
}
\usage{
mqtls_NG
}
\description{
Author: A. Gallois
Title: A comprehensive study of metabolite genetics reveals strong pleiotropy and heterogeneity across time and contex
Journal: Nature Communications https://doi.org/10.1038/s41467-019-12703-7
}
\keyword{datasets}
