% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Muscle_Skeletal.R
\docType{data}
\name{Muscle_Skeletal}
\alias{Muscle_Skeletal}
\title{Matrix of expression data of Muscle...Skeletal from gtex}
\format{
A [[matrix]] object
}
\source{
[Muscle_Skeletal](https://gtexportal.org/home/datasets)
}
\usage{
Muscle_Skeletal
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
