% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nerve_Tibial.R
\docType{data}
\name{Nerve_Tibial}
\alias{Nerve_Tibial}
\title{Matrix of expression data of Nerve...Tibial from gtex}
\format{
A [[matrix]] object
}
\source{
[Nerve_Tibial](https://gtexportal.org/home/datasets)
}
\usage{
Nerve_Tibial
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
