% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovary.R
\docType{data}
\name{Ovary}
\alias{Ovary}
\title{Matrix of expression data of Ovary from gtex}
\format{
A [[matrix]] object
}
\source{
[Ovary](https://gtexportal.org/home/datasets)
}
\usage{
Ovary
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
