% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pancreas.R
\docType{data}
\name{Pancreas}
\alias{Pancreas}
\title{Matrix of expression data of Pancreas from gtex}
\format{
A [[matrix]] object
}
\source{
[Pancreas](https://gtexportal.org/home/datasets)
}
\usage{
Pancreas
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
