% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pituitary.R
\docType{data}
\name{Pituitary}
\alias{Pituitary}
\title{Matrix of expression data of Pituitary from gtex}
\format{
A [[matrix]] object
}
\source{
[Pituitary](https://gtexportal.org/home/datasets)
}
\usage{
Pituitary
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
