% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtls.R
\docType{data}
\name{pqtls}
\alias{pqtls}
\title{Data frame of pQTLs}
\format{
A [[data.frame]] object
}
\source{
[pqtls](see description)
}
\usage{
pqtls
}
\description{
Yao, C., et al., Genome‐wide mapping of plasma protein QTLs identifies putatively causal genes and pathways for cardiovascular disease. Nature Communications, 2018. 9(1): p. 3268
}
\details{
Wu, L., et al., Variation and genetic control of protein abundance in humans. Nature, 2013. 499: p. 79.

Suhre, K., et al., Connecting genetic risk to disease end points through the human blood plasma proteome. Nature communications, 2017. 8: p. 14357-14357.

Gudmundsdottir, V et al Circulating Protein Signatures and Causal Candidates for Type 2 Diabetes Diabetes 2020 Aug; 69(8): 1843-1853.
}
\keyword{datasets}
