% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prostate.R
\docType{data}
\name{Prostate}
\alias{Prostate}
\title{Matrix of expression data of Prostate from gtex}
\format{
A [[matrix]] object
}
\source{
[Prostate](https://gtexportal.org/home/datasets)
}
\usage{
Prostate
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
