% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinationRate.R
\docType{data}
\name{recombinationRate}
\alias{recombinationRate}
\title{List of data frames with recombination rates for chromosomes 1-22}
\format{
A [[list]] object
}
\source{
[recombinationRate](ftp://ftp.ncbi.nlm.nih.gov/hapmap/)
}
\usage{
recombinationRate
}
\description{
HapMap project
}
\keyword{datasets}
