% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skin_Not_Sun_Exposed_Suprapubic.R
\docType{data}
\name{Skin_Not_Sun_Exposed_Suprapubic}
\alias{Skin_Not_Sun_Exposed_Suprapubic}
\title{Matrix of expression data of Skin...Not.Sun.Exposed..Suprapubic. from gtex}
\format{
A [[matrix]] object
}
\source{
[Skin_Not_Sun_Exposed_Suprapubic](https://gtexportal.org/home/datasets)
}
\usage{
Skin_Not_Sun_Exposed_Suprapubic
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
