% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skin_Sun_Exposed_Lower_leg.R
\docType{data}
\name{Skin_Sun_Exposed_Lower_leg}
\alias{Skin_Sun_Exposed_Lower_leg}
\title{Matrix of expression data of Skin...Sun.Exposed..Lower.leg. from gtex}
\format{
A [[matrix]] object
}
\source{
[Skin_Sun_Exposed_Lower_leg](https://gtexportal.org/home/datasets)
}
\usage{
Skin_Sun_Exposed_Lower_leg
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
