% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small_Intestine_Terminal_Ileum.R
\docType{data}
\name{Small_Intestine_Terminal_Ileum}
\alias{Small_Intestine_Terminal_Ileum}
\title{Matrix of expression data of Small.Intestine...Terminal.Ileum from gtex}
\format{
A [[matrix]] object
}
\source{
[Small_Intestine_Terminal_Ileum](https://gtexportal.org/home/datasets)
}
\usage{
Small_Intestine_Terminal_Ileum
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
