% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spleen.R
\docType{data}
\name{Spleen}
\alias{Spleen}
\title{Matrix of expression data of Spleen from gtex}
\format{
A [[matrix]] object
}
\source{
[Spleen](https://gtexportal.org/home/datasets)
}
\usage{
Spleen
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
