% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stomach.R
\docType{data}
\name{Stomach}
\alias{Stomach}
\title{Matrix of expression data of Stomach from gtex}
\format{
A [[matrix]] object
}
\source{
[Stomach](https://gtexportal.org/home/datasets)
}
\usage{
Stomach
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
