% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testis.R
\docType{data}
\name{Testis}
\alias{Testis}
\title{Matrix of expression data of Testis from gtex}
\format{
A [[matrix]] object
}
\source{
[Testis](https://gtexportal.org/home/datasets)
}
\usage{
Testis
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
