% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thyroid.R
\docType{data}
\name{Thyroid}
\alias{Thyroid}
\title{Matrix of expression data of Thyroid from gtex}
\format{
A [[matrix]] object
}
\source{
[Thyroid](https://gtexportal.org/home/datasets)
}
\usage{
Thyroid
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
