% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptionFactorsAll.R
\docType{data}
\name{TranscriptionFactorsAll}
\alias{TranscriptionFactorsAll}
\title{GRanges object of transcription factor binding sites}
\format{
A [[GRanges]] object
}
\source{
[TranscriptionFactorsAll](https://genome.ucsc.edu/ENCODE/downloads.html)
}
\usage{
TranscriptionFactorsAll
}
\description{
ENCODE project
}
\keyword{datasets}
