% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uterus.R
\docType{data}
\name{Uterus}
\alias{Uterus}
\title{Matrix of expression data of Uterus from gtex}
\format{
A [[matrix]] object
}
\source{
[Uterus](https://gtexportal.org/home/datasets)
}
\usage{
Uterus
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
