% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vagina.R
\docType{data}
\name{Vagina}
\alias{Vagina}
\title{Matrix of expression data of Vagina from gtex}
\format{
A [[matrix]] object
}
\source{
[Vagina](https://gtexportal.org/home/datasets)
}
\usage{
Vagina
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
