% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Whole_Blood.R
\docType{data}
\name{Whole_Blood}
\alias{Whole_Blood}
\title{Matrix of expression data of Whole.Blood from gtex}
\format{
A [[matrix]] object
}
\source{
[Whole_Blood](https://gtexportal.org/home/datasets)
}
\usage{
Whole_Blood
}
\description{
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
