% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTLs.R
\docType{data}
\name{pQTLs}
\alias{pQTLs}
\title{Data frame of pQTLs}
\format{
A [[data.frame]] object
}
\source{
[pQTLs](see description)
}
\usage{
pQTLs
}
\description{
Yao, C., et al., Genome‐wide mapping of plasma protein QTLs identifies putatively causal genes and pathways for cardiovascular disease. Nature Communications, 2018. 9(1): p. 3268
}
\details{
Wu, L., et al., Variation and genetic control of protein abundance in humans. Nature, 2013. 499: p. 79.

Suhre, K., et al., Connecting genetic risk to disease end points through the human blood plasma proteome. Nature communications, 2017. 8: p. 14357-14357.

Carayol, J., et al., Protein quantitative trait locus study in obesity during weight-loss identifies a leptin regulator. Nature Communications, 2017. 8(1): p. 2084.
}
\keyword{datasets}
